% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.r
\name{check_version}
\alias{check_version}
\title{Functions to support installation and loading of CRAN and Bioconductor packages in R}
\usage{
check_version(pkg_name, min_version)
}
\arguments{
\item{x}{a package_name}
}
\description{
Functions to support installation and loading of CRAN and Bioconductor packages in R
}
