% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.r
\name{CheckAvailableEnsemblOrganisms}
\alias{CheckAvailableEnsemblOrganisms}
\title{Check Ensembl for available Organisms}
\usage{
CheckAvailableEnsemblOrganisms(
  url = "ftp://ftp.ensembl.org/pub/",
  release = NULL
)
}
\arguments{
\item{release}{}
}
\description{
Check Ensembl for available Organisms
}
