% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.r
\name{CheckAvailableEnsemblReleases}
\alias{CheckAvailableEnsemblReleases}
\title{Check Ensembl for available releases}
\usage{
CheckAvailableEnsemblReleases(url = "ftp://ftp.ensembl.org/pub/")
}
\arguments{
\item{url}{}
}
\description{
Check Ensembl for available releases
}
