% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GO_plots.r
\name{clean_GO_output}
\alias{clean_GO_output}
\title{Cleans Gene Ontology outputs and combines with ReviGO reduced data}
\usage{
clean_GO_output(results, revigo)
}
\arguments{
\item{revigo}{}
}
\value{
df
}
\description{
Cleans Gene Ontology outputs and combines with ReviGO reduced data
}
