% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_to_TPM.r
\name{counts_to_TPM}
\alias{counts_to_TPM}
\title{Title}
\usage{
counts_to_TPM(dds, annot)
}
\arguments{
\item{dds}{}

\item{annot}{full annotation from GTF model (created with prepare_annotation.r)}
}
\description{
Title
}
\examples{
load("D:/Box Sync/Projects/CHM/RNA-seq/analysis/CHMvsWT/deseq2-HF-CHMvsRefWT-t_all-WALDp0.05/deseq2-HF-CHMvsRefWT_Rdata.rdata")
annot <- read.csv("D:/annotation.v104_danio_rerio/annotationDanio_rerio.GRCz11.104.gtf.annotation.full.from.GTF.csv")
a<-counts_to_TPM(dds, annot)
}
