% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GO_charts.r
\name{create_GO_barchart}
\alias{create_GO_barchart}
\title{Create Gene Ontology (GO) Bar Charts}
\usage{
create_GO_barchart(df, ont, fill)
}
\arguments{
\item{df}{data.frame out Webgestalt output}

\item{ont}{Ontology to be assessed (MF/BP/CC etc)}

\item{fill}{RColorBrewer colour palette for example "Reds" for further information check http://applied-r.com/rcolorbrewer-palettes/}
}
\value{
ggplot2 object of the bar chart
}
\description{
Create Gene Ontology (GO) Bar Charts
}
