% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.r
\name{DownloadEnsemblGTF}
\alias{DownloadEnsemblGTF}
\title{Download Ensembl GTF}
\usage{
DownloadEnsemblGTF(
  version = NULL,
  url = "ftp.ensembl.org/pub/",
  temp_dir = NULL,
  organism = NULL,
  data_type = "genes"
)
}
\arguments{
\item{data_type}{}
}
\description{
Download Ensembl GTF
}
