% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KEGG_genes.r
\name{get_KEGG_genes}
\alias{get_KEGG_genes}
\title{Get genes for a specific KEGG pathway}
\usage{
get_KEGG_genes(pathwayquery)
}
\arguments{
\item{pathwayquery}{search term keyword (human)}
}
\description{
returns geneList a list of all human and zebrafish gene IDs related to pathway KEGG with keyword used.
To get a complete list of pathways to subsequently search, use the keyword 'ALL' .
}
