% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_barchart_plots.r
\name{go_barchart_plots}
\alias{go_barchart_plots}
\title{Go make those barchart plots for Webgestalt output}
\usage{
go_barchart_plots(fileList, filePath)
}
\arguments{
\item{fileList}{to loop over}

\item{filePath}{to find the files at}
}
\description{
Go make those barchart plots for Webgestalt output
}
