% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.r
\name{install_package}
\alias{install_package}
\title{Functions to support installation and loading of CRAN and Bioconductor packages in R}
\usage{
install_package(x)
}
\arguments{
\item{x}{a package_name}
}
\description{
Functions to support installation and loading of CRAN and Bioconductor packages in R
}
