% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.r
\name{load_pack}
\alias{load_pack}
\title{Functions to support installation and loading of CRAN and Bioconductor packages in R}
\usage{
load_pack(x, warn_conflicts = T)
}
\arguments{
\item{x}{a package_name}
}
\description{
Functions to support installation and loading of CRAN and Bioconductor packages in R
}
