% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GO_plots.r
\name{make_GO_plots}
\alias{make_GO_plots}
\title{Creates Gene Ontology plots for Enrichment}
\usage{
make_GO_plots(df, ont, fill)
}
\arguments{
\item{df}{combined output from clean_GO_output}

\item{ont}{Ontology type BP/CC/MF/KEGG etc}

\item{fill}{Plot colour}
}
\value{
p
}
\description{
Creates Gene Ontology plots for Enrichment
}
