% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_organoid_goi.R
\name{plot_organoid_goi}
\alias{plot_organoid_goi}
\title{Plot Organoid Expression GOI: Gene of Interest}
\usage{
plot_organoid_goi(goi, timepoint, conditions_to_keep)
}
\arguments{
\item{goi}{\itemize{
\item gene of interest as string
}}

\item{timepoint}{\itemize{
\item day time point of data, options 'All', 'd00', 'd20', 'd35'
}}

\item{conditions_to_keep}{\itemize{
\item list of conditions to keep in the plot ('All' or others)
}}
}
\description{
Plot Organoid Expression GOI: Gene of Interest
}
\examples{
plot_organoid_goi("CRB1", "All", "PX,WT,AN,BM,UM,MY")


}
