% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA_forpub.r
\name{plot_PCA_forpub}
\alias{plot_PCA_forpub}
\title{Plot PCA for publication format}
\usage{
plot_PCA_forpub(rlogdata, samplesToRemove = "NONE", output, outFormat = "svg")
}
\arguments{
\item{rlogdata}{\itemize{
\item object of transformed data from DESeq2
}}

\item{samplesToRemove}{\itemize{
\item list object of samples to remove from dataset, default="NONE")
}}

\item{output}{\itemize{
\item filename of the output in cwd
}}

\item{outFormat}{\itemize{
\item output format of the plot (default = SVG)
}}
}
\description{
Plot PCA for publication format
}
