% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Sample2Sample.r
\name{plot_Sample2Sample}
\alias{plot_Sample2Sample}
\title{Plot sample to sample distance heatmap for samples in DDS object}
\usage{
plot_Sample2Sample(rld, outputFilename, outputFormat = "svg")
}
\arguments{
\item{rld}{\itemize{
\item transformed dds object from DESeq2
}}

\item{outputFilename}{\itemize{
\item filename of output image
}}

\item{outputFormat}{\itemize{
\item format of output image
}}
}
\description{
Plot sample to sample distance heatmap for samples in DDS object
}
