% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.r
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare annotation for RNA-seq data analysis from Ensembl}
\usage{
prepare_annotation(species, ver, loc)
}
\arguments{
\item{species}{valid Ensembl species name}

\item{ver}{version number}

\item{loc}{location of where to store annotation files}
}
\description{
Prepare annotation for RNA-seq data analysis from Ensembl
}
