% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_GO_plots.r
\name{clean_GO_output}
\alias{clean_GO_output}
\title{Cleans Gene Ontology outputs and combines with ReviGO reduced data}
\usage{
clean_GO_output(results, revigo)
}
\arguments{
\item{results}{GO over representation results data.frame}

\item{revigo}{ReviGO reduced output data.frame

Both data.frames should contain first column with GO terms}
}
\value{
data.frame
}
\description{
Cleans Gene Ontology outputs and combines with ReviGO reduced data
}
